% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramids.R
\name{pyramids}
\alias{pyramids}
\title{Generate artwork structured around a grid of pyramids viewed from above.}
\usage{
pyramids(
  n_rows = 5,
  n_cols = 10,
  gap_size = 0.1,
  buffer = 0.1,
  neutrals = NULL,
  accents = NULL
)
}
\arguments{
\item{n_rows}{number of rows of pyramids in plot.}

\item{n_cols}{number of columns of pyramids in plot.}

\item{gap_size}{relative spacing between pyramids.}

\item{buffer}{relative size of internal buffer; prevents peaks falling too close to pyramid edges.}

\item{neutrals}{(optional) vector of neutral colours.}

\item{accents}{(optional) vector of accent colours.}
}
\value{
ggplot object
}
\description{
Generate artwork structured around a grid of pyramids viewed from above.
}
