% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zv_palettes.R
\name{scale_color_zv_d}
\alias{scale_color_zv_d}
\title{Change colours in ggplot for discrete variables}
\usage{
scale_color_zv_d(palette_name, ...)
}
\arguments{
\item{palette_name}{Name of palette.}

\item{...}{other arguments to be passed to \code{scale_colour_manual}}
}
\description{
Change colours in ggplot for discrete variables
}
