% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zv_palettes.R
\name{scale_colour_zv_c}
\alias{scale_colour_zv_c}
\title{Change colours in ggplot for continuous variables}
\usage{
scale_colour_zv_c(palette_name, ...)
}
\arguments{
\item{palette_name}{Name of Palette.}

\item{...}{other arguments to be passed to \code{scale_colour_gradientn}}
}
\description{
Change colours in ggplot for continuous variables
}
