% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zv_palettes.R
\name{scale_fill_zv_c}
\alias{scale_fill_zv_c}
\title{Change fill in ggplot for continuous variables}
\usage{
scale_fill_zv_c(palette_name = "seq_grey", ...)
}
\arguments{
\item{palette_name}{Name of Palette.}

\item{...}{other arguments to be passed to \code{scale_fill_gradientn}}
}
\description{
Change fill in ggplot for continuous variables
}
