% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zv_palettes.R
\name{scale_fill_zv_d}
\alias{scale_fill_zv_d}
\title{Change fill in ggplot for discrete variables}
\usage{
scale_fill_zv_d(palette_name, ...)
}
\arguments{
\item{palette_name}{Name of Palette.}

\item{...}{other arguments to be passed to \code{scale_fill_manual}}
}
\description{
Change fill in ggplot for discrete variables
}
