% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_definitions.R
\docType{data}
\name{zv_black}
\alias{zv_black}
\title{Define global colours}
\format{
An object of class \code{character} of length 1.
}
\usage{
zv_black
}
\description{
Define global colours
}
\keyword{datasets}
