% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_definitions.R
\docType{data}
\name{zv_colours}
\alias{zv_colours}
\title{List of colours.}
\format{
An object of class \code{list} of length 7.
}
\usage{
zv_colours
}
\description{
Complete list of palettes.
Use names(lu_colours) to return all palette names.
}
\keyword{datasets}
