% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zv_palettes.R
\name{zv_palettes}
\alias{zv_palettes}
\title{Generate a palette from zv_colours}
\usage{
zv_palettes(palette_name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{palette_name}{Name of generated palette. Character string.}

\item{n}{Number of desired colours. If number of requested colours is
beyond the scope of the palette, colours are automatically interpolated.}

\item{type}{The type of palette. One of c("discrete", "continuous").}
}
\value{
A vector of colours.
}
\description{
Generates colour palette from zvplot colour definitions.
}
