% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_zv.R
\name{theme_zv}
\alias{theme_zv}
\title{A place-holder function for my ggplot theme}
\usage{
theme_zv(string = "This is a placeholder function")
}
\arguments{
\item{string}{a character string to return}
}
\value{
\code{string}
}
\description{
A place-holder function for my ggplot theme
}
\examples{
theme_zv()
theme_zv("Hello, World!")
}
